requirements.txt
azure-functions
f5-sdk-python